JS addition and update to fix apple menus
Safari (and all iOS and macOS browsers) are forced to use apple web API. This API's handling of mouse events is flawed (a holdover from pre mobile days) and so the original CSS menus were not working.

Added JS to improve on the raw CSS menus where available and to begin making the site more friendly for screen readers.